fix: consistenly use u32 for paging (#3)
* adjust some docs

* shorten some lenghty macro

* use link to top-level axum crate

* Make page use all u32

* fmt

* fix typo

+++ examples/kitchen-sink/examples/api_pagination.rs
@@ -78,8 +78,8 @@ pub struct Book {
 #[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
 pub struct PaginatedBook {
     pub page: u32,
-    pub page_size: u64,
-    pub count: u64,
+    pub page_size: u32,
+    pub count: u32,
     pub data: Vec<Book>,
 }
